requirements.txt
python-docx
pyyaml
openai
anthropic
tqdm